feat: show results

+++ lib/socket-event-handlers/user-leave-room.ts
@@ -25,17 +25,18 @@ export const handleUserLeaveRoom = async (
     .from(participants)
     .where(eq(participants.roomId, roomId));
 
-  if (!remainingParticipants.length) {
+  const [room] = await db.select().from(rooms).where(eq(rooms.id, roomId));
+
+  // Check if host has left the room
+  if (room.creatorId === userId) {
+    await db.update(rooms).set({ creatorId: remainingParticipants[0].userId });
+  }
+
+  // Check if room is empty and inactive
+  if (!remainingParticipants.length && !room.isActiveRoom) {
     await db.delete(rooms).where(eq(rooms.id, roomId));
     socket.broadcast.emit("updateRoomList");
     return;
-  }
-
-  // Check if host has left the room
-  const [room] = await db.select().from(rooms).where(eq(rooms.id, roomId));
-
-  if (room.creatorId === userId) {
-    await db.update(rooms).set({ creatorId: remainingParticipants[0].userId });
   }
 
   socket.to(`room-${roomId}`).emit("updateRoom", { roomId });
